// lib/pov/types/core.ts
export type PoVMetadata = {
  customer: string;
  teamSize: string;
  successCriteria: string;
  technicalRequirements: string;
  [key: string]: any;
};